+++ src/Exceptions/Exception.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Swis\JsonApi\Client\Exceptions;
+
+interface Exception
+{
+}