+++ src/Exceptions/ValidationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Swis\JsonApi\Client\Exceptions;
+
+class ValidationException extends \InvalidArgumentException implements Exception
+{
+}