+++ src/Links.php
@@ -112,8 +112,8 @@ class Links implements ArrayAccess, Arrayable, Jsonable, JsonSerializable
     public function toArray()
     {
         return array_map(
-            function (Link $link) {
-                return $link->toArray();
+            static function (?Link $link) {
+                return $link ? $link->toArray() : null;
             },
             $this->links
         );